web: score neighbours at 1.0 when a tile has solar
When a tile is submitted as having solar panels, set all versions of the
neighbouring tiles to a score of 1.0. This lets us quickly review all
tiles for installations that span multiple tiles. Reviewing the
surrounding area provides negative examples that likely look similar to
the positive ones which should be useful to avoid overfitting to buiding
styles. Similarly, reviewing older versions of the same tiles often lets
us provide examples of the same building before and after installation.