Allow accepting predicted score as ground truth
For areas a negative answer makes sense, score it as 0, but a generic
positive answer doesn't (what should the area be in that case?). Instead
add a new option to accept the prediction as ground truth. This probably
leads to a new, longer, loop of predict, add to OSM if needed, ignore
predicted score, and then later run an import of scores from OSM,
accepting the scores that come from the newly drawn features.