download: download neighbours even if not scheduled
If we found something new at one positions it is likely that the
surrounding tiles have updates as well, even if they are below the limit
where we would normally try again. If it turns out they don't have
anything new after all we won't search any further in that direction, so
this won't cause any significant load increase.